Refactor input models into appropriate subfolders
