+++ src/electionguard_cli/cli_models/__init__.py
@@ -1,6 +1,5 @@
 from electionguard_cli.cli_models import e2e_build_election_results
 from electionguard_cli.cli_models import e2e_decrypt_results
-from electionguard_cli.cli_models import e2e_inputs
 from electionguard_cli.cli_models import e2e_submit_results
 
 from electionguard_cli.cli_models.e2e_build_election_results import (
@@ -9,24 +8,15 @@ from electionguard_cli.cli_models.e2e_build_election_results import (
 from electionguard_cli.cli_models.e2e_decrypt_results import (
     E2eDecryptResults,
 )
-from electionguard_cli.cli_models.e2e_inputs import (
-    CliElectionInputsBase,
-    E2eInputs,
-    ImportBallotInputs,
-)
 from electionguard_cli.cli_models.e2e_submit_results import (
     E2eSubmitResults,
 )
 
 __all__ = [
     "BuildElectionResults",
-    "CliElectionInputsBase",
     "E2eDecryptResults",
-    "E2eInputs",
     "E2eSubmitResults",
-    "ImportBallotInputs",
     "e2e_build_election_results",
     "e2e_decrypt_results",
-    "e2e_inputs",
     "e2e_submit_results",
 ]
