+++ src/electionguard_cli/steps/shared/election_record_step.py
@@ -4,7 +4,7 @@ from tempfile import TemporaryDirectory
 from click import echo
 from electionguard_cli.cli_models import BuildElectionResults, E2eSubmitResults
 from electionguard_cli.cli_models.e2e_decrypt_results import E2eDecryptResults
-from electionguard_cli.cli_models.e2e_inputs import E2eInputs
+from electionguard_cli.cli_models.e2e.e2e_inputs import E2eInputs
 from electionguard.constants import get_constants
 
 from electionguard_tools.helpers.export import export_record
